+++ lib/src/ui/splitter.dart
@@ -3,6 +3,10 @@
 
 part of dart_web_toolkit_ui;
 
+/**
+ * Abstract Splitter for SplitLayoutPanelExample. Implemented as Splitter and 
+ * VSplitter.
+ */
 abstract class Splitter extends Widget {
   Widget target;